requirements/base.txt
# Core extraction
pymupdf>=1.24.0
pdfplumber>=0.11.0
ftfy>=6.2.0

# Embedding + vector DB
# NOTE on Apple Silicon with x86_64 Anaconda (Rosetta): torch max is 2.2.2 via pip.
# This requires sentence-transformers<3 and transformers<5.
# On Linux/native arm64 conda: remove the upper bounds.
torch>=2.2.2
sentence-transformers>=2.7,<3    # <3 required for torch 2.2.2 on x86 Mac
transformers>=4.41,<5            # <5 avoids breaking changes in 5.x
numpy>=1.26.0,<2.0               # torch 2.2.2 requires numpy <2
qdrant-client>=1.9.0
einops>=0.8.0              # required by nomic-embed-text-v1.5

# Clustering
# NOTE: on Apple Silicon (M1/M2), install numba+llvmlite via conda first:
#   conda install -c conda-forge numba llvmlite
# Then pip install hdbscan umap-learn normally.
# Alternatively, scikit-learn >= 1.3 has a built-in HDBSCAN (no numba needed).
umap-learn>=0.5.6
scikit-learn>=1.4.0
# hdbscan is optional — sklearn's HDBSCAN is used automatically if available
hdbscan>=0.8.38.post1

# LLM / API
anthropic>=0.30.0

# Data / utilities
pydantic>=2.7.0
loguru>=0.7.2
tqdm>=4.66.0
rich>=13.7.0
pandas>=2.2.0

# Reranking
# (sentence-transformers also covers cross-encoder)
